+++ modules/Users/Search.inc.php
@@ -32,7 +32,7 @@ if ( empty( $_REQUEST['search_modfunc'] ) )
 
 			echo '<form name="search" id="search" action="' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] .
 				'&modfunc=' . $_REQUEST['modfunc'] .
-				'&search_modfunc=list&next_modname=' . $_REQUEST['next_modname'] .
+				'&search_modfunc=list' .
 				'&advanced=' . issetVal( $_REQUEST['advanced'], '' ) .
 				issetVal( $extra['action'], '' )  ) . '" method="GET">';
 
@@ -162,11 +162,6 @@ if ( empty( $_REQUEST['search_modfunc'] ) )
 //if ( $_REQUEST['search_modfunc']=== 'list')
 else
 {
-	if ( empty( $_REQUEST['next_modname'] ) )
-	{
-		$_REQUEST['next_modname'] = 'Users/User.php';
-	}
-
 	if ( empty( $extra['NoSearchTerms'] ) )
 	{
 		$_ROSARIO['SearchTerms'] = issetVal( $_ROSARIO['SearchTerms'] );
@@ -225,7 +220,7 @@ else
 		];
 	}
 
-	$name_link['FULL_NAME']['link'] = 'Modules.php?modname=' . $_REQUEST['next_modname'];
+	$name_link['FULL_NAME']['link'] = 'Modules.php?modname=' . $_REQUEST['modname'];
 	$name_link['FULL_NAME']['variables'] = [ 'staff_id' => 'STAFF_ID' ];
 
 	if ( isset( $extra['link'] )
@@ -331,38 +326,6 @@ else
 
 			// Unset search modfunc & redirect URL.
 			RedirectURL( 'search_modfunc' );
-		}
-
-		if ( $_REQUEST['modname'] != $_REQUEST['next_modname'] )
-		{
-			$modname = $_REQUEST['next_modname'];
-
-			if ( mb_strpos( $modname, '?' ) )
-			{
-				$modname = mb_substr( $_REQUEST['next_modname'], 0, mb_strpos( $_REQUEST['next_modname'], '?' ) );
-			}
-
-			if ( mb_strpos( $modname, '&' ) )
-			{
-				$modname = mb_substr( $_REQUEST['next_modname'], 0, mb_strpos( $_REQUEST['next_modname'], '&' ) );
-			}
-
-			if ( ! empty( $_REQUEST['modname'] ) )
-			{
-				$_REQUEST['modname'] = $modname;
-			}
-
-			//FJ security fix, cf http://www.securiteam.com/securitynews/6S02U1P6BI.html
-
-			if ( mb_substr( $modname, -4, 4 ) != '.php' || mb_strpos( $modname, '..' ) !== false || ! is_file( 'modules/' . $modname ) )
-			{
-				require_once 'ProgramFunctions/HackingLog.fnc.php';
-				HackingLog();
-			}
-			else
-			{
-				require_once 'modules/' . $modname;
-			}
 		}
 	}
 	else